Poker_analyze/requirements.txt
# 核心功能依賴
anthropic==0.3.0
tensorflow-macos==2.12.0
tensorflow-metal==0.8.0
numpy==1.23.5
pillow==9.5.0

# 截圖與監控依賴
pynput==1.7.6
pyautogui==0.9.54

# 後端服務依賴
websockets==10.4
python-dotenv==1.0.0
requests==2.31.0

# 開發工具依賴
black==23.3.0
pylint==2.17.3